[refact] move workers out of trainers
