+++ verl/workers/reward_model/megatron/reward_model.py
@@ -28,7 +28,7 @@ from verl.utils.megatron.pipeline_parallel import (compute_transformers_input_sh
 from verl import DataProto
 from verl.utils.torch_functional import logprobs_from_logits, broadcast_dict_tensor, split_dict_tensor_into_batches
 from verl.utils.torch_dtypes import PrecisionType
-from verl.trainer.workers.reward_model.base import BasePPORewardModel
+from verl.workers.reward_model.base import BasePPORewardModel
 from verl.utils.megatron import sequence_parallel as sp_utils
 from megatron.core import parallel_state as mpu
 from megatron.core.pipeline_parallel import get_forward_backward_func
